Refactor/merge partner projects into a single component
